mail-client/evolution: Misc improvements
Make sure that e-d-s is built with gtk USE otherwise a cryptic error is
shown to users instead of dialog that allows user to take action and
continue.

Also add dconf to dependencies. This is used for saving the settings
to disk and therefore useful for people who use evo outside GNOME.
Without dconf the 'memory' backend is used which is cleared after
application exits.