eclass: add XDG eclass to replace fdo-mime
fdo-mime eclass is just a collection of helpers that is most often
forgotten and/or not properly used in ebuilds having .desktop files or
shared mime infos.

On top of that, most packages where missing the implicit dependency on
the tools used as can be seen in bug #208047.

Finally, most packages complying with XDG specifications nowadays expect
base directories to be present at various stages of package build as can
be seen in the many eclasses implementing handling for this (cmake, gnome2,
gstreamer, kde4-base, mono and qt4) so provide it through a simple phase
function.