gnome-extra/gnome-characters: remove, available in ::gentoo
Main tree version removes all of the test stuff, as FEATURES=test and
IUSE=test is separate; thus no point in having the dependencies in place,
if tests aren't going to work (for correctness we'd have to package.use.mask
tests for the package as well, and instead have a
RESTRICT="!test? ( test )").
Also removes unnecessary vala_src_prepare - gnome-characters is written
in javascript, not vala.